app/backend/connector/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.11.0
asyncio==4.0.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
fastapi==0.118.0
Flask==3.1.2
flask-cors==6.0.1
flask-localtunnel==1.0.7
frozenlist==1.7.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.6.4
mysql-connector-python==9.4.0
nest-asyncio==1.6.0
packaging==25.0
propcache==0.4.0
psycopg2-binary==2.9.10
py-localtunnel==1.0.3
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
requests==2.32.5
rich==14.1.0
setuptools==80.9.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
Werkzeug==3.1.3
yarl==1.20.1
